+++ src/test/kotlin/com/github/hanseter/json/editor/base/EditorTestBase.kt
@@ -2,11 +2,9 @@ package com.github.hanseter.json.editor.base
 
 import com.github.hanseter.json.editor.JsonPropertiesEditor
 import com.github.hanseter.json.editor.ui.ControlTreeItemData
-import com.github.hanseter.json.editor.ui.CustomTreeTableView
-import com.sun.javafx.robot.FXRobot
 import javafx.scene.Scene
 import javafx.scene.control.TreeTablePosition
-import javafx.scene.input.KeyCode
+import javafx.scene.control.TreeTableView
 import javafx.stage.Stage
 import org.json.JSONObject
 import org.junit.jupiter.api.extension.ExtendWith
@@ -19,12 +17,10 @@ import org.testfx.util.WaitForAsyncUtils
 open class EditorTestBase {
 
     lateinit var editor: JsonPropertiesEditor
-    lateinit var robot: FxRobot
 
     @Start
     open fun start(stage: Stage) {
         editor = JsonPropertiesEditor()
-        robot = FxRobot()
         stage.scene = Scene(editor, 1500.0, 1000.0)
         stage.show()
     }
@@ -41,11 +37,11 @@ open class EditorTestBase {
         displayData(data,schema)
     }
 
-    fun getValueCells(): MutableSet<JsonPropertiesEditor.ValueCell> {
+    fun getValueCells(robot: FxRobot): MutableSet<JsonPropertiesEditor.ValueCell> {
         return robot.lookup(".cell.control-cell").queryAll()
     }
     fun getCellValue(selectedCell: TreeTablePosition<out Any?, *>?) =
         (selectedCell?.treeItem?.value as ControlTreeItemData).typeControl.model.value
 
-    fun getTable()=robot.lookup(".tree-table-view").queryAs(CustomTreeTableView::class.java)
+    fun getTable(robot: FxRobot)=robot.lookup(".tree-table-view").queryAs(TreeTableView::class.java)
 }
